ListOfList complete. need comments. Now converted in h file. 25-11
